React File structure updated
