--- PatrolCam.US-main/next-app/next-app/src/app/landing/layout.js
@@ -1,11 +1,0 @@
-//layout for landing page
-
-export default function LandingLayout({ children }) {
-    return (
-        <html lang="en">
-            <body>
-                {children}
-            </body>
-        </html>
-    )
-}
