--- PatrolCam.US-main/next-app/next-app/src/app/landing/page.js
@@ -1,25 +1,0 @@
-//landing page for PatrolCam 
-import Head from 'next/head';
-
-
-export default function LandingPage() {
-  return (
-    <>
-      <Head>
-        <meta charSet="UTF-8"></meta>
-        <meta name="viewport" content="width=device-width, initial-scale=1.0"></meta>
-        <title>PatrolCam</title>
-      </Head>
-      {/* this is where navbar is coded */}
-
-      {/* First part of the actual website users will see */}
-      <div className="container">
-          <h1>PatrolCam</h1>
-      </div>
-
-      {/* Start of the features page */}
-
-      {/* Footer */}
-    </>
-  );
-}
