+++ PatrolCam.US-main/next-app/next-app/src/app/page.jsx
@@ -1,0 +1,14 @@
+//landing page for PatrolCam 
+
+export default function Page() {
+  return (
+    <>
+      {/* First part of the actual website users will see */}
+      <div className="container">
+          <h1>PatrolCam</h1>
+      </div>
+
+      {/* Start of the features page */}
+    </>
+  );
+}
