+++ PatrolCam.US-main/next-app/next-app/src/app/ui/footer.jsx
@@ -1,0 +1,14 @@
+// footer used throughout all of Patrol Cam
+
+export default function Footer() {
+    return (
+        <>
+            <footer>
+                {/* style for footer using Tailwind CSS */}
+                <div> 
+                    <p>© 2024 PatrolCam. All rights reserved.</p>
+                </div>
+            </footer>
+        </>
+    );
+}
