+++ PatrolCam.US-main/next-app/next-app/src/app/ui/navbar.jsx
@@ -1,0 +1,19 @@
+// Navbar used throughout all of Patrol cam 
+// 'use client'
+// import Link from 'next/link';
+
+// //placeholder navbar until the backend is set up
+
+// export default function Navbar() {
+//     return (
+//         <div>
+//             <nav>
+//                 <ul>
+//                     <li><Link href="/featureIntro"></Link>Features</li>
+//                     <li><Link href="/demo"></Link>Demo</li>
+//                     <li><Link href="/contactSection">Contact Us</Link></li>
+//                 </ul>
+//             </nav>
+//         </div>
+//     );
+// }
